Project name is changed & white __init__.py
Issue #16 is done.
This project don't use only IMAP protocol, Both IMAP and SMTP
Therefore, I change my project name Thanks
