+++ .env
@@ -1,5 +1,5 @@
-export PROJECT_HOME=/home/likemilk/workspace/pi-imap-transmission
-export PROJECT_NAME=pi-imap-transmission
+export PROJECT_HOME=/home/likemilk/workspace/pi-email-transmission
+export PROJECT_NAME=pi-email-transmission
 
 #LOG_LEVE
 #   develop
