+++ init.sh
@@ -10,7 +10,7 @@ pyenv install 3.4.3
 
 pyenv virtualenv 3.4.3 $PROJECT_NAME
 
-ln -s `pyenv root`/versions/$PROJECT_NAME $PROJECT_HOME/$PROJECT_NAME
+# ln -s `pyenv root`/versions/$PROJECT_NAME $PROJECT_HOME/$PROJECT_NAME
 
 pyenv shell $PROJECT_NAME
 pip install --upgrade pip
