+++ pet/src/__init__.py
@@ -1,0 +1,2 @@
+from pet.src.service import PETService
+from pet.src.thread import PETThread
