+++ pet/src/controller/mail_controller.py
@@ -1,24 +1,14 @@
-import sys
 import imaplib
 from smtplib import SMTP_SSL
+from smtplib import SMTPException
 import email
-import base64
-import json
 
-from pit.utils.logger import DJLogger
-if sys.version_info < (3,):
-    from urllib2 import Request
-    from urllib2 import HTTPError
-    from urllib2 import urlopen
-else:
-    from urllib.request import Request
-    from urllib.error import HTTPError
-    from urllib.request import urlopen as urlopen
+from pet.utils import Logger
 
-logger = DJLogger.getLogger()
+logger = Logger.getLogger()
 
 
-class IMAPController:
+class MailController:
     email_format = '''From: %s
 To: %s
 MIME-Version: 1.0
@@ -60,8 +50,9 @@ Subject: %s
                 email
             )
             server.quit()
-        except Exception as e:
+        except SMTPException as e:
             logger.debug(str(e))
+            logger.debug(e)
             logger.debug(str(e.__dict__))
 
     def check(self):
@@ -112,7 +103,7 @@ Subject: %s
             .replace("{sizeWhenDone}", str(torrent_info['sizeWhenDone']))\
             .replace("{files}", "</br>".join(file_list))
 
-        email_form = IMAPController.email_format % (
+        email_form = MailController.email_format % (
             self.config['service']['email']['user'],
             self.config['service']['email']['user'],
             email_form['subject'].replace('{subject}', torrent_info['name']),
@@ -147,7 +138,7 @@ Subject: %s
         content = email_form['content']
         content = content.replace("{files}", "</br>".join(file_list))
         print("149149149: ", "</br>".join(file_list))
-        email_form = IMAPController.email_format % (
+        email_form = MailController.email_format % (
             self.config['service']['email']['user'],
             self.config['service']['email']['user'],
             email_form['subject'],
@@ -174,7 +165,7 @@ Subject: %s
             .replace("{sizeWhenDone}", str(torrent_info['sizeWhenDone']))\
             .replace("{files}", "</br>".join(file_list))
 
-        email_form = IMAPController.email_format % (
+        email_form = MailController.email_format % (
             self.config['service']['email']['user'],
             self.config['service']['email']['user'],
             email_form['subject'].replace('{subject}', torrent_info['name']),
@@ -202,113 +193,13 @@ Subject: %s
             .replace("{sizeWhenDone}", str(torrent_info['sizeWhenDone']))\
             .replace("{files}", "</br>".join(file_list))
 
-        email_form = IMAPController.email_format % (
+        email_form = MailController.email_format % (
             self.config['service']['email']['user'],
             self.config['service']['email']['user'],
             email_form['subject'].replace('{subject}', torrent_info['name']),
             content
         )
         self.__send_email__(email_form)
-
-
-class TransmissionController:
-    def __init__(self, config):
-        self.__config = config
-        self.__session_id = None
-        self.__auth = base64.encodestring(
-            (
-                '%s:%s' %
-                (
-                    self.__config['service']['transmission']['user'],
-                    self.__config['service']['transmission']['password']
-                )
-            ).encode()
-        ).decode().replace('\n', '')
-        self.__get_session__()
-
-    def __request__(self, method, arguments=None):
-        message = {"method": method, "arguments": arguments}
-        message = json.dumps(message).encode("utf-8")
-
-        req = Request(
-            "%s/transmission/rpc"
-            % (self.__config['service']['transmission']['host']),
-            message
-        )
-
-        req.add_header("Content-Type", "application/json")
-        req.add_header("Authorization", "Basic %s" % self.__auth)
-
-        res = None
-        for i in range(3):
-            try:
-                if self.__session_id:
-                    req.add_header(
-                        "X-Transmission-Session-Id",
-                        self.__session_id
-                    )
-                res = urlopen(req)
-            except HTTPError as e:
-                if e.code == 409:
-                    self.__session_id = e.headers['X-Transmission-Session-Id']
-
-        res = json.loads(res.read().decode('utf-8'))
-        return res
-
-    def __get_session__(self):
-        self.__request__('session-get')
-
-    def __torrent_info__(self, torrent_id):
-        method = 'torrent-get'
-        field = self.__config['field'][method]
-        field['ids'] = [torrent_id, ]
-        print('203 : ', field)
-        return self.__request__(method, field)['arguments']['torrents'][0]
-
-    def check(self):
-        complete_torrent = []
-        method = 'torrent-get'
-        res = self.__request__(method, self.__config['field'][method])
-        if (res['result'] != 'success'):
-            raise HTTPError('transmission has occured an error at torrent-get')
-
-        for torrent in res['arguments']['torrents']:
-            if(torrent['percentDone'] is 1):
-                complete_torrent.append(torrent)
-        return complete_torrent
-
-    def add(self, payload):
-        method = 'torrent-add'
-        res = self.__request__(method, {"metainfo": payload})
-        if(res['result'] != 'success'):
-            return None
-        else:
-            torrent = res['arguments'].get('torrent-added', None)
-            torrent_id = 0
-            if torrent is not None:
-                torrent_id = torrent['id']
-            else:
-                torrent_id = res['arguments']['torrent-duplicate']['id']
-
-            return self.__torrent_info__(torrent_id)
-
-    def delete(self, torrent):
-        method = 'torrent-remove'
-        torrent_info = self.__torrent_info__(torrent['id'])
-        res = self.__request__(method, {"ids": torrent['id']})
-        if(res['result'] != 'success'):
-            return None
-        else:
-            return torrent_info
-
-    def pause(self, torrent):
-        method = 'torrent-stop'
-        torrent_info = self.__torrent_info__(torrent['id'])
-        res = self.__request__(method, {"ids": torrent['id']})
-        if(res['result'] != 'success'):
-            return None
-        else:
-            return torrent_info
 
 
 '''
