+++ pet/test/test_logger.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
-from pit.utils.logger import DJLogger
-from pit.__main__ import main
+from pet.utils.logger import DJLogger
+from pet.__main__ import main
 
 '''
 
@@ -8,7 +8,7 @@ from pit.__main__ import main
 '''
 
 
-class TestPIT(TestCase):
+class TestPET(TestCase):
     def test_is_logger(self):
         logger = DJLogger.getLogger()
         logger.debug('So curious!')
