+++ pet/utils/__init__.py
@@ -1,0 +1,2 @@
+from pet.utils.logger import Logger
+from pet.utils.config import PETConfig
