+++ pet/utils/config/context.py
@@ -1,0 +1,2 @@
+class PETContext:
+    TEXT = 'THIS IS TEXT'
