--- pi-imap-transmission
@@ -1,1 +1,0 @@
-/home/likemilk/.pyenv/versions/pi-imap-transmission
