--- pit/__init__.py
@@ -1,8 +1,0 @@
-from pit.utils.logger import DJLogger
-
-print ('init.py')
-
-
-'''
-
-'''
